+++ src/lib.rs
@@ -7,6 +7,7 @@ pub(crate) mod gpu;
 pub mod group;
 pub(crate) mod identity;
 pub(crate) mod io;
+pub(crate) mod keys;
 pub(crate) mod mem;
 pub(crate) mod net;
 pub mod once;